Remove event hashing
It was a huge burden of complexity for no gain.
A Project is a materialised view over a stream of events.
Deriving hashes from it and then expecting them to forever hold or even
have value, was incorrect because the materialised view is going to
change semi-regularly to support new features. There are infinite
possible ways of representing the same project data; tightly-coupling
hash partitions and algorithms to them was the wrong idea. If the idea
of hashes were to be reintroduced it would be better to have a universal
(as in minimal, canonical, not-further factorisable) representation of
sections of the project. Even that would have evolutionary concerns and
asks a significant complexity budget for nearly no value. What is the
value that hashes were providing? Confidence that you generate the same
content? Why? For data integrity or prevention against corruption, then
hashes over the events themselves instead of a Project seems more
appropriate.

Regardless, the current implementation isn't desirable.
Into the bin it goes, sadly.