Remove all prefetching
Chrome doesn't handle it well. It doesn't use an origin during
prefetching and then it uses the cached rejection when you click a
project. Funny thing is that if you hit reload it works, but you always
have to hit reload. So fucking stupid.

Just rip it all out and simplify my life.